Move auto-collections files to their own package
